fix: clippy warnings

+++ src/position.rs
@@ -40,7 +40,7 @@ impl Position {
     pub fn move_left(&self, by: usize) -> Position {
         Position {
             line: self.line,
-            column: self.column.saturating_sub(by as usize),
+            column: self.column.saturating_sub(by),
         }
     }
